feat(ui): update handling of inProgess, do not allow cnet process when processing
